Tests for HRS and CDR, custom exceptions for HRS

+++ HRS/src/main/java/rom/hrs/repository/CallPricingRepository.java
@@ -7,6 +7,6 @@ import rom.hrs.entity.CallPricing;
 import java.util.List;
 
 @Repository
-public interface CallPricingRepository extends JpaRepository<CallPricing, Integer> {
-    List<CallPricing> findByTariffId(Integer tariffId);
+public interface CallPricingRepository extends JpaRepository<CallPricing, Long> {
+    List<CallPricing> findByTariffId(Long tariffId);
 }
